move report listeners to correct file, and force UTM tags into segments data source
